services/autoscan/requirements.txt
pymongo
python-dotenv
# awscli
# boto3